+++ src/main/resource/config.py
@@ -5,10 +5,27 @@ class Config(object):
     # Определяет, включен ли режим отладки
     # В случае если включен, flask будет показывать
     # подробную отладочную информацию. Если выключен -
-    # - 500 ошибку без какой либо дополнительной информации.
+    # - 500 ошибку без какой-либо дополнительной информации.
     DEBUG = False
+    ROOT_DIR = os.path.dirname(os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__)))))
     MODEL_PATH = "data/models/"
-    VOICE_PATH = "data/wavs/"
+    AUDIO_FILES_DIR_PATH = os.path.join(ROOT_DIR, "data/wavs/")
+    AUDIO_CONVERSION_INPUT_DIR = os.path.join(ROOT_DIR, "data/ogg/")
+    AUDIO_CONVERSION_OUTPUT_EXT = "wav"
+    WORDS_FILE_PATH = os.path.join(ROOT_DIR, "data/words.txt")
+    WORDS_FILE_ENCODING = "utf-8"
+    PHONEMES_FILE_PATH = os.path.join(ROOT_DIR, "data/phonemes.txt")
+    RECOGNIZE_LANGUAGE = "ru-RU"
+    PHONEMIZE_LANGUAGE = "ru"
+    PHONEMIZE_BACKEND = "espeak"  # Install ESPEAK NG as msi from github releases
+
+    AUDIO_GENERATION_LEARNING_RATE = 0.001
+    AUDIO_GENERATION_BATCH_SIZE = 16
+    AUDIO_GENERATION_NUM_EPOCHS = 50
+    AUDIO_GENERATION_NUM_MELS = 80  # number of mel spectrogram bins
+    AUDIO_GENERATION_VOCAB_SIZE = 80
+    AUDIO_GENERATION_CHECKPOINT_DIR_PATH = os.path.join(ROOT_DIR, "data/checkpoints/")
+
     SERVER_PORT = 5000
     CSRF_ENABLED = True
     # Случайный ключ, которые будет исползоваться для подписи данных, например cookies.